requirements.txt
asgiref==3.3.1
Django==3.1.3
django-cleanup==5.2.0
django-cors-headers==2.5.2
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.4
django-utils-six==2.0
djangorestframework==3.11.1
Pillow==8.0.1
pytz==2020.4
sqlparse==0.4.1
whitenoise==5.3.0

